pyproject.toml
[project]
name = "crm-platform-aws"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3 (>=1.38.21,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "moto (>=5.1.4,<6.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.0"
pytest = "^8.3.5"

